+++ Moloco/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.10.1.0
+* Certified with Moloco SDK 3.10.1.
+* Update ad display failed error code.
+
 ## 3.10.0.0
 * Certified with Moloco SDK 3.10.0.